Added GitHub workflow to build and test Go code.
This PR adds a GitHub workflow to build and test the Go code. See [this
GitHub pull request][1] for a similar PR.

=== Accessing Private Repos ===

The trickiest part of getting this workflow working was allowing the
weaver-gke repo to access the private weaver repo. To get this to work,
we generate an RSA key pair called a [deploy key][4]. A deploy key is
a lot the SSH keys you use personally to authenticate with GitHub but
it's for a machine instead of a person.

After creating the key pair, we upload the public key to the weaver repo
and the private key as a secret to the weaver-gke repo. In the GitHub
workflow, we ssh-add the private key and set GOPRIVATE appropriately to
access the weaver repo.

See [2] and [3] for more information on this process.

=== Details ===

- I decreased the go.mod version from go 1.20 to 1.19 since the code
  works with go 1.19.
- I converted protoc.sh to use bash instead of sh to make the script
  work with GitHub actions.
- I added a .golangci.yaml lint config to make the linter work.

[1]: https://github.com/ServiceWeaver/weaver/pull/2
[2]: https://zerokspot.com/weblog/2022/11/18/private-go-modules-on-github/
[3]: https://github.com/marketplace/actions/webfactory-ssh-agent
[4]: https://docs.github.com/en/developers/overview/managing-deploy-keys