Advance GKE traffic fractions if the app is receiving no traffic. (#21)
In particular, suppose that the application version exports a number
of listeners, with hostnames H1, ..., Hn. Currently, we examine the
traffic for each hostname, and determine whether the actual traffic
fraction for a given application version matches the expected traffic.
Only if all of the hostnames H1, ..., Hn get the expected traffic do
we advance the traffic fraction for the app version.

The issue is that some hostname Hi may not receive any traffic
*across all app versions* that export that hostname. In that case,
we cannot compute the traffic fraction for our target application
version.

In this pull request, we change the above behavior to treat Hi
as receiving the expected traffic fraction. This will help roll out
new application versions even when the application itself receives
no traffic.