Check if a metric is available before querying it in GKE.
Prior to this change, if the controller queried the http_count
metric that hasn't yet been created (e.g., no traffic sent to
the app), the query would fail with a random error type.

This change modifies this behavior to (1) detect whether the
metric actually exists, and (2) return empty counts if it
doesn't.

Other changes:
  * Add more permissions to the computer service account. Again,
    this may only be needed for internal cloud deployments, but
    these permissions are actually needed.