Fix the fleet registration bug.
The bug was caused by the newer gcloud version, which now requires
(and supports) the "--location" flag when creating a new memberhsip
binding. These bindings no longer default to the "global" location,
resulting in regional fleet memberships.

While at it, check the gcloud version before starting a deployment,
and encourage users to update the gcloud to the latest version.