Updated to weaver v0.13.0.
The only notable change is the deletion of managertest. weavertest does
not work with the new weaver API (at least I couldn't think of an easy
way to get it working). I decided to delete the tests since they were a
bit hacky anyway and were fragile. The manager is already tested in
other ways.

DO_NOT_SUBMIT: Update go.mod to v0.13.0 when it is released.