Use access token to authenticate with the Google Cloud.
When using GitHub actions or other CI/CD platform, the user may
not be able to run `gcloud init` interactively. The recommended
login for those kinds of deployments is to get an access token
from Google Cloud, store it as a secret and use it for deployments.
This change adds support for using such tokens for deployment, either
by passing an "access_token" flag to the `weaver-gke` commands, or
by setting the CLOUDSDK_AUTH_ACCESS_TOKEN variable when invoking
the commands.

Other changes:
  * Unify flag handling for the cloud flags across all commands.
  * Remove project and account fields from the `GKEConfig` proto.
  * Minor cleanups.