Fix a race in creating CA certificate/key between tool and controller.
This race was introduced by a recent change where we changed the
order of calls to `Controller` and `PrepareRollout` while deploying.
This introduced a known race where the controller and the tool attempt
to create and store the shared CA certificate key on the local file system.
In turn, this caused the tool and the controller to use different
CA certificates, resulting in a TLS error between them.

This change restores the old behavior of the tool generating the CA
certificate before starting the controller.

Other changes:
  * Add a simple test for certificates (unrelated to this fix).