Use the Pod watcher to detect the set of remote component replicas.
This change mimics the related change in weaver-kube where the babysitter
watches the activated component's pods and returns the replicas therein.
The GKE deployer was previously consulting the assigner for an updated
set of replicas, which is a complicated flow fraught with possibilities
of errors.

Moreover, an upcoming weaver change will see weavelets performing health
checks themselves, obsoleting the current GKE mechanism which has the
assigner performing health checks on the babysitters.

TODO: Change the assigner to only manage routing information for routed
components.