Use the same hostname for all versions of the echo app.
Currently, GCP has a rather small default limit on the number of SSL
certificates it can maintain. Additionally, we don't have a good way
of deleting certificates when they are no longer needed. For this
reason, we temporarily reduce the number of certificates by reusing
the same hostname in integration tests, namely, "echo.serviceweaver.dev".

We will implement correct certificate deletion in a future change.