Make the rollout knob GKE deployer specific
Right now the `rollout` knob is global, and it applies to all deployers.
However, each deployer has its own rollout story. E.g., single, multi -
no rollout story, SSH - no rollout story, kube - up to the DevOps people
how to do it.

The rollout knob makes sense only for the GKE deployer, where we have a
custom rollout strategy, so it makes sense to make the knob a GKE
deployer specific knob. However, if in the future we'll decide to add a
system component that can compute rollout strategies and it can be
applied to many deployers, we can move it back.

I will remove the `rollout` knob from the weaver repo in a followup PR.