Rewrite and simplify the assigner code.
In particular, we now store all of the assigner state in a series
of ReplicaSetState entries in the store, and find application
and application-version subsets of those using the `store.List`
operation.

Additionally, all Pods in a ReplicaSet are discovered, health-
checked, and load-inspected entirely by the assigner, periodically.
This simplifies the state update at the Assigner significantly.

Other changes:
  * Unify health-checking for routed and non-routed components
    back at the assigner.
  * A small number of cleanups throughout the code.