Fix gcloud error handling.
The latest gcloud API update returns a gRPC error type with an Unknown
status code, so we shouldn't take a gRPC status presence as a sign that
the error has the gRPC type.

Instead, we now check all available status codes to see if any of
them is of type "not found".

This fixes an issue found by Valentin when deploying to a brand
new project.