Don't create special config cluster.
In the current implementation, the GKE deployer creates a cluster for
each region where the application is being deployed. Also, it creates a
config cluster in "us-central1" where the controller is being deployed.

This is not great, because for each user project, we create an
additional cluster which translates into cloud costs for the user.

This PR deploys the controller in one of the clusters where applications
that are part of the same user project are being deployer.

If this is the first time the user deploys an app in their project, the
controller is deployed in the first region where the app is being
deployed. However, if the controller is already running in some cluster,
it will run there forever.

The caveat with this approach is that if the user decides some day to
not deploy anything anymore in the region where the controller has been
running, we'll still run the controller in that region.