Enable user to specify min replicas for components
Right now we will start one pod for each component (colocation group).
We will rely on the multi dim pod autoscaler to scale up / down to more
pods. However, for some users, they want the guarantee that each
component is running in at least a minimum number of pods (which can be
greater than 1).

This PR enables the user to specify the minimum number of replicas for
components in the config. Note that this value is unique and it will be
enforced across all the components.