Enforce at least #replicas for each component to run.
Currently, the GKE deployer will run at least one replica for each
component. However, based on feedback from the customers, sometimes they
want a guarantee on the minimum number of running replicas for each
component, for availability reasons.

This PR introduces a `minreplicas` knob in the config that enables the
user to configure the minimum number of running pods for each component.