Purge the deprecated config cluster
We changed the way we deploy the Service Weaver config cluster; i.e., we
deploy the controller in one of the application clusters instead.

However, for old deployments, the user might still have the config
cluster (that runs in us-central1) around.

This PR updates the purge command to delete the us-central1 config
cluster as well (if exists).