Fix the cache key for the babysitter http clients (#135)
In #134 we changed the code to create a single http client for each
babysitter. However, we keep track of existing http clients for each
babysitter indexed by replica set. But, the pods in a replica set can
change based on load, hence we need to create a http client per replica
set and babysitter address.