Allow users to specify custom image
In the current implementation, we use "ubuntu:rolling" as the default
image to build the application. However, we don't provide to the user a way
to configure the base image for the application; e.g., the application might
require some custom packages that are not installed by default in
"ubuntu:rolling"; or the user wants to add their own custom code.

There are two options to allow the user to customize the docker image:
1) expose the dockerfile template to the user (or some form of it); however, this is not great because the gke app config becomes more verbose, we need to provide a subtemplate to be configurable, and also the user might get confused.
2) allow the user to specify a base image in the gke app config. With
   this approach, the user can optionally specify a base image.
   Otherwise, we're gonna use "ubuntu:rolling". This PR implements this
   approach.

Note that with 2) the user has to have an image that is pushed to a
container registry, and accessible from the user project. Also, the user
has to specify the full name of the image in the config, based on which
container registry is it pushed.

Error handling: I don't think we need to check whether the image exists,
and/or if the user specify the name of the image correctly. Cloudbuild
will throw an error if there are any issues building the app on top of
the custom base image provided by the user.