Configure names for private listeners
Before this PR, the user is able to specify a hostname for the public
listeners. However, for the private listeners, we compute an internal
name by default. This is not great, especially if you want to change the
names of the internal listeners (e.g., deploy in different
environments).

This PR changes the way the users can configure the listeners:
1) If the user specify a listener config, it has to specify whether the
   listener is public by setting `is_public` knob explicitly.
2) If `is_public` knob is not set, the listener will be private by
   default.
3) For public listeners, the user is required to specify the `hostname`.
4) For private listeners, `hostname` is optional; if not set, we'll
   create a default name:
   <listener_name>.<region>.serviceweaver.internal; if set, the name
   will be:
   <hostname>.<listener_name>.<region>.serviceweaver.internal