Replace the removed BackendConfig with HealthCheckPolicy resource.
GKE has (silently and not backward-compatibly) removed the BackendConfig
as the mechanism to specify health-check policies. The new mechanism
is based on the new HealthCheckPolicy resource. This change migrates
to the new mechanism.

A bit of difficulty is that we can no longer attach the same health-check
resource to multiple ServiceImports. Moreover, the new health-check
resource requires the Pod port to be specified. This change addresses
both of these issues.