Delete autoscalers when rolling out new versions (#147)
In the current implementation, when a new version of the app is fully
rolled out, we delete the kubernetes resources associated with the old
app (except the autoscalers). This is because we're not setting the
app and the version labels in the kubernetes manifest for the multi
dimensional autoscaler. However, this is not easy. Somehow if we
manually set the labels in the template, the labels are completely
ignored once the template is converted to an unstructured object.

In this PR we add the app and the version labels to the kubernetes
manifest for the multi pod autoscaler. Also, we delete the autoscaler
resources for old deployments.

Note that we also have a kill command that has the same issue. This PR
should fix it.